# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Baccarat prediction dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface SessionUser {
  id: number;
  username: string;
  /** ISO 8601 datetime when account expires */
  expiresAt: string;
}

export interface AdminCredentials {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  ok: boolean;
}

export interface GameInputData {
  /** B, P, or T */
  value: string;
}

export interface WindowSetting {
  /** Rolling window: 8, 12, or 16 */
  window: number;
}

export interface RegimeExpert {
  predCount: number;
  /** Weighted win rate (0-1) */
  wwr: number;
  /** Raw win rate (0-1) */
  rawWr: number;
}

export interface RegimeState {
  /** WARMING_UP | TRACKING | SPLIT */
  status: string;
  /**
     * B | P | null (wait)
     * @nullable
     */
  decision: string | null;
  /**
     * supreme | syndicate | null
     * @nullable
     */
  expert: string | null;
  /** NONE | LOW | MED | HIGH */
  confidence: string;
  isSplit: boolean;
  gap: number;
  bothAgree: boolean;
  /** @nullable */
  bothAgreeSide?: string | null;
  regimeAge: number;
  switchCount: number;
  justSwitched: boolean;
  isLocked?: boolean;
  supreme: RegimeExpert;
  syndicate: RegimeExpert;
  window: number;
}

export interface GameSnapshot {
  handCount: number;
  /** Array of B, P, T outcomes entered so far */
  history: string[];
  regime: RegimeState;
}

export interface UserAccount {
  id: number;
  username: string;
  /** ISO 8601 datetime */
  expiresAt: string;
  /** ISO 8601 datetime */
  createdAt: string;
  isOnline: boolean;
  /**
     * ISO 8601 datetime or null
     * @nullable
     */
  lastLoginAt: string | null;
}

export interface UserAccountInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 6 */
  password: string;
  /** ISO 8601 datetime */
  expiresAt: string;
}

export interface UserAccountUpdate {
  /** @minLength 1 */
  username?: string;
  /** @minLength 6 */
  password?: string;
  /** ISO 8601 datetime */
  expiresAt?: string;
}