# lib/api-zod/src/generated/types/gameSnapshot.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Baccarat prediction dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { RegimeState } from './regimeState';

export interface GameSnapshot {
  handCount: number;
  /** Array of B, P, T outcomes entered so far */
  history: string[];
  regime: RegimeState;
}