# lib/api-zod/src/generated/types/regimeExpert.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Baccarat prediction dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RegimeExpert {
  predCount: number;
  /** Weighted win rate (0-1) */
  wwr: number;
  /** Raw win rate (0-1) */
  rawWr: number;
}