# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Baccarat prediction dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionUser {
  id: number;
  username: string;
  /** ISO 8601 datetime when account expires */
  expiresAt: string;
}