# lib/api-zod/src/generated/types/userAccount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Baccarat prediction dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UserAccount {
  id: number;
  username: string;
  /** ISO 8601 datetime */
  expiresAt: string;
  /** ISO 8601 datetime */
  createdAt: string;
  isOnline: boolean;
  /**
     * ISO 8601 datetime or null
     * @nullable
     */
  lastLoginAt: string | null;
}