# lib/api-zod/src/generated/types/userAccountUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Baccarat prediction dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UserAccountUpdate {
  /** @minLength 1 */
  username?: string;
  /** @minLength 6 */
  password?: string;
  /** ISO 8601 datetime */
  expiresAt?: string;
}